+++ ArrhythmicBattles/Util/Screen.cs
@@ -6,7 +6,7 @@ namespace ArrhythmicBattles.Util;
 
 public abstract class Screen : IDisposable
 {
-    public abstract Vector2i Position { get; set; }
+    public abstract Vector2d Position { get; set; }
 
     public abstract void Render(Renderer renderer, int layerId, MatrixStack matrixStack, CameraData cameraData);
     public abstract void Update(UpdateArgs args);
